Updated to work with PHP7 and OpenRVS
I made multiple changes to allow working with PHP7. Mainly mysql to mysqli, $HTTP_GET_VARS to $_GET, and $HTTP_COOKIE_VARS to $_COOKIE.

I also made changes to allow this to work with a modified version of OpenRVS. Ubi servers shutdown, and OpenRVS does not currently send an Ubi name. I have modified OpenRVS (branch to come) to push an UBINAME_PASS and then modified the code here to parse it. This will allow a player to provide a unique name and key. This name can then be associated to their nicknames, and stats can properly be recorded to individual players. I also stripped out the ubi ladder to only display offline, and modified the UBI name to display only the name and not the password.

Finally, I made some minor bug changes and added some code to prevent PHP warnings.