Changes to allow Unique ID to be checked against Server List
This is the code that allows the server to check the provided user id and password against the information stored in the server INI. If the user does not exist, we create a new entry. If the user exists and the passwords match, we allow that user to use the name provided. If the passwords do not match, we append a unique string to the username and notify the user.

Modified post and server .php to use the names sent from the server and not to  look for the user_pass anymore as this is done by the game server.